pyproject.toml
[project]
name = "devsecops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "jinja2==2.11.2",
    "psycopg2-binary>=2.9.11",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
